SONARJAVA-6827: Refactor S9357 to eliminate duplication with S1604
Extract shared detection logic into AbstractAnonymousClassToLambdaCheck
base class used by both AnonymousClassOnFunctionalInterfaceCheck (S9357)
and AnonymousClassShouldBeLambdaCheck (S1604).

Change S9357 scope from "All" to "Tests" to avoid double-reporting
on main code where S1604 already applies. Move S9357 test samples
to test source path accordingly.
