+++ sonar-java-plugin/src/main/resources/org/sonar/l10n/java/rules/java/S9357.json
@@ -14,7 +14,7 @@
   "defaultSeverity": "Major",
   "ruleSpecification": "RSPEC-9357",
   "sqKey": "S9357",
-  "scope": "All",
+  "scope": "Tests",
   "quickfix": "unknown",
   "code": {
     "impacts": {
